refactor(tui): split ui.rs into per-pane modules
Closes #2266
